requirements.txt
astropy
bokeh
Django
requests
django-environ
djangorestframework
